fix merged in changes
